Resolving test errors

+++ pkg/service/query_test.go
@@ -65,7 +65,7 @@ func TestQueryServiceNodePortWithIngress(t *testing.T) {
 	assert.Len(t, svc.Spec.Ports, 1)
 	assert.Equal(t, int32(16686), svc.Spec.Ports[0].Port)
 	assert.Equal(t, "http-query", svc.Spec.Ports[0].Name)
-	assert.Equal(t, 0, svc.Spec.Ports[0].NodePort)
+	assert.Equal(t, int32(0), svc.Spec.Ports[0].NodePort)
 	assert.Equal(t, intstr.FromInt(16686), svc.Spec.Ports[0].TargetPort)
 	assert.Equal(t, svc.Spec.Type, corev1.ServiceTypeNodePort) // make sure we get a NodePort service
 }
@@ -97,9 +97,9 @@ func TestQueryServiceSpecifiedNodePortWithIngress(t *testing.T) {
 
 	assert.Equal(t, "testqueryservicespecifiednodeportwithingress-query", svc.ObjectMeta.Name)
 	assert.Len(t, svc.Spec.Ports, 1)
-	assert.Equal(t, 16686, svc.Spec.Ports[0].Port)
+	assert.Equal(t, int32(16686), svc.Spec.Ports[0].Port)
 	assert.Equal(t, "http-query", svc.Spec.Ports[0].Name)
-	assert.Equal(t, 32767, svc.Spec.Ports[0].NodePort) // make sure we get the same NodePort as set above
+	assert.Equal(t, int32(32767), svc.Spec.Ports[0].NodePort) // make sure we get the same NodePort as set above
 	assert.Equal(t, intstr.FromInt(16686), svc.Spec.Ports[0].TargetPort)
 	assert.Equal(t, svc.Spec.Type, corev1.ServiceTypeNodePort)
 }
